requirements.txt
# ===== Core ML =====
torch==2.5.1
transformers==4.48.0
tokenizers==0.21.0

# ===== PEFT & Quantization =====
peft==0.7.1
bitsandbytes==0.45.0
accelerate==1.2.1

# ===== Training =====
trl==0.13.0
datasets==2.21.0

# ===== Tokenizer deps =====
sentencepiece==0.2.0
protobuf==5.29.0

# ===== SQL Processing =====
sqlparse==0.5.3

# ===== Data & Visualization =====
pandas==2.2.3
numpy==1.26.4
matplotlib==3.10.0
seaborn==0.13.2

# ===== Experiment Tracking (optional) =====
# wandb==0.19.0

# ===== Utilities =====
python-dotenv==1.0.1
tqdm==4.67.0
pyyaml==6.0.2
